+++ Resteurent/WSRestaurant/Reposetory/OrderReposetory.cs
@@ -16,9 +16,15 @@ namespace WSRestaurant
 
         public bool delete(string id)
         {
-            string sql = $@"DELETE FROM Orders WHERE OrderId=@OrderId";
+            string sql = $@"DELETE DishOrder WHERE OrderId=@OrderId";
             this.dbContext.AddParameter("@OrderId", id);
-            return this.dbContext.Delete(sql);
+            if (this.dbContext.Delete(sql))
+            {
+                sql = $@"DELETE FROM Orders WHERE OrderId=@OrderId";
+                this.dbContext.AddParameter("@OrderId", id);
+                return this.delete(sql);
+            }
+            else throw new Exception("return false");
         }
 
         public List<Orders> getAll()
@@ -46,12 +52,37 @@ namespace WSRestaurant
                 return this.modelFactory.createOrderObject.CreateModel(dataReader);
             }
         }
+
+        public Orders getByCustomer(string customerId)
+        {
+            string sql = "SELECT FROM Orders WHERE CustomerId = @CustomerId";
+            this.dbContext.AddParameter("@CustomerId", customerId);
+            using (IDataReader dataReader = this.dbContext.Read(sql))
+            {
+                dataReader.Read();
+                return this.modelFactory.createOrderObject.CreateModel(dataReader);
+            }
+        }
         public bool update(Orders model)
         {
             string sql = $@"UPDATE Orders SET CustomerId = @CustomerId, OrderDate = @OrderDate WHERE OrderId == @OrderId";
             this.dbContext.AddParameter("@CustomerId", model.Customer.Id);
             this.dbContext.AddParameter("@OrderDate", model.OrderDate.ToString());
             return this.dbContext.Update(sql);
+        }
+
+        public bool deleteByCustomer(string customerId)
+        {
+            Orders order = this.getByCustomer(customerId);
+            string sql = $@"DELETE FROM Orders WHERE CustomerId=@CustomerId";
+            this.dbContext.AddParameter("@CustomerId", customerId);
+            if (this.dbContext.Delete(sql))
+            {
+                sql = $@"DELETE FROM DishOrder WHERE OrderId=@OrderId";
+                this.dbContext.AddParameter("@DishId", order.Id);
+                return this.dbContext.Delete(sql);
+            }
+            else throw new Exception("return false");
         }
     }
 }
